fixed text alignment with RTL languages, shows an initial intro prompt the first time you log in, fixed bug with gender selection control

+++ Record/Record/EAFAudioPlayer.h
@@ -23,7 +23,7 @@
 
 @property NSString *url;
 @property NSArray *audioPaths;
-@property int currentIndex;
+@property unsigned long currentIndex;
 
 - (IBAction)playRefAudio;
 - (IBAction)stopAudio;
